requirements.txt
numpy>=2.3.3
torch>=2.9.0+cu130
plyfile>=1.1.2
trimesh>=4.8.3
scikit-image>=0.25.2
tqdm>=4.67.1
PyYAML>=6.0.3
scipy>=1.16.2
munch>=4.0.0
pytorch-lightning>=2.5.5
setuptools>=70.2.0
numba>=0.62.1
torchvision>=0.24.0+cu130